Reorganized the code into separate files.
