+++ wideq/__init__.py
@@ -4,12 +4,15 @@ import requests
 from urllib.parse import urljoin, urlencode, urlparse, parse_qs
 import uuid
 import base64
-import json
 import hashlib
 import hmac
 import datetime
-from collections import namedtuple
-import enum
+
+# These imports are for backwards compatiability with previous versions.
+from .errors import (
+    APIError, MonitorError, NotConnectedError, NotLoggedInError, TokenError)
+from .device import DeviceInfo, ModelInfo
+from .device.ac import ACDevice, ACFanSpeed, ACMode, ACOp, ACStatus  # noqa
 
 __version__ = '1.0.1'
 
@@ -61,45 +64,6 @@ def get_list(obj, key):
         return val
     else:
         return [val]
-
-
-class APIError(Exception):
-    """An error reported by the API."""
-
-    def __init__(self, code, message):
-        self.code = code
-        self.message = message
-
-
-class NotLoggedInError(APIError):
-    """The session is not valid or expired."""
-
-    def __init__(self):
-        pass
-
-
-class NotConnectedError(APIError):
-    """The service can't contact the specified device."""
-
-    def __init__(self):
-        pass
-
-
-class TokenError(APIError):
-    """An authentication token was rejected."""
-
-    def __init__(self):
-        pass
-
-
-class MonitorError(APIError):
-    """Monitoring a device failed, possibly because the monitoring
-    session failed and needs to be restarted.
-    """
-
-    def __init__(self, device_id, code):
-        self.device_id = device_id
-        self.code = code
 
 
 def lgedm_post(url, data=None, access_token=None, session_id=None):
@@ -400,59 +364,6 @@ class Session(object):
         return res['returnData']
 
 
-class Monitor(object):
-    """A monitoring task for a device.
-
-    This task is robust to some API-level failures. If the monitoring
-    task expires, it attempts to start a new one automatically. This
-    makes one `Monitor` object suitable for long-term monitoring.
-    """
-
-    def __init__(self, session, device_id):
-        self.session = session
-        self.device_id = device_id
-
-    def start(self):
-        self.work_id = self.session.monitor_start(self.device_id)
-
-    def stop(self):
-        self.session.monitor_stop(self.device_id, self.work_id)
-
-    def poll(self):
-        """Get the current status data (a bytestring) or None if the
-        device is not yet ready.
-        """
-
-        try:
-            return self.session.monitor_poll(self.device_id, self.work_id)
-        except MonitorError:
-            # Try to restart the task.
-            self.stop()
-            self.start()
-            return None
-
-    @staticmethod
-    def decode_json(data):
-        """Decode a bytestring that encodes JSON status data."""
-
-        return json.loads(data.decode('utf8'))
-
-    def poll_json(self):
-        """For devices where status is reported via JSON data, get the
-        decoded status result (or None if status is not available).
-        """
-
-        data = self.poll()
-        return self.decode_json(data) if data else None
-
-    def __enter__(self):
-        self.start()
-        return self
-
-    def __exit__(self, type, value, tb):
-        self.stop()
-
-
 class Client(object):
     """A higher-level API wrapper that provides a session more easily
     and allows serialization of state.
@@ -609,460 +520,3 @@ class Client(object):
         if url not in self._model_info:
             self._model_info[url] = device.load_model_info()
         return ModelInfo(self._model_info[url])
-
-
-class DeviceType(enum.Enum):
-    """The category of device."""
-
-    REFRIGERATOR = 101
-    KIMCHI_REFRIGERATOR = 102
-    WATER_PURIFIER = 103
-    WASHER = 201
-    DRYER = 202
-    STYLER = 203
-    DISHWASHER = 204
-    OVEN = 301
-    MICROWAVE = 302
-    COOKTOP = 303
-    HOOD = 304
-    AC = 401  # Includes heat pumps, etc., possibly all HVAC devices.
-    AIR_PURIFIER = 402
-    DEHUMIDIFIER = 403
-    ROBOT_KING = 501  # Robotic vacuum cleaner?
-    ARCH = 1001
-    MISSG = 3001
-    SENSOR = 3002
-    SOLAR_SENSOR = 3102
-    IOT_LIGHTING = 3003
-    IOT_MOTION_SENSOR = 3004
-    IOT_SMART_PLUG = 3005
-    IOT_DUST_SENSOR = 3006
-    EMS_AIR_STATION = 4001
-    AIR_SENSOR = 4003
-
-
-class DeviceInfo(object):
-    """Details about a user's device.
-
-    This is populated from a JSON dictionary provided by the API.
-    """
-
-    def __init__(self, data):
-        self.data = data
-
-    @property
-    def model_id(self):
-        return self.data['modelNm']
-
-    @property
-    def id(self):
-        return self.data['deviceId']
-
-    @property
-    def model_info_url(self):
-        return self.data['modelJsonUrl']
-
-    @property
-    def name(self):
-        return self.data['alias']
-
-    @property
-    def type(self):
-        """The kind of device, as a `DeviceType` value."""
-
-        return DeviceType(self.data['deviceType'])
-
-    def load_model_info(self):
-        """Load JSON data describing the model's capabilities.
-        """
-        return requests.get(self.model_info_url).json()
-
-
-EnumValue = namedtuple('EnumValue', ['options'])
-RangeValue = namedtuple('RangeValue', ['min', 'max', 'step'])
-
-
-class ModelInfo(object):
-    """A description of a device model's capabilities.
-    """
-
-    def __init__(self, data):
-        self.data = data
-
-    def value(self, name):
-        """Look up information about a value.
-
-        Return either an `EnumValue` or a `RangeValue`.
-        """
-        d = self.data['Value'][name]
-        if d['type'] in ('Enum', 'enum'):
-            return EnumValue(d['option'])
-        elif d['type'] == 'Range':
-            return RangeValue(
-                d['option']['min'], d['option']['max'], d['option']['step']
-            )
-        else:
-            assert False, "unsupported value type {}".format(d['type'])
-
-    def default(self, name):
-        """Get the default value, if it exists, for a given value.
-        """
-
-        return self.data['Value'][name]['default']
-
-    def enum_value(self, key, name):
-        """Look up the encoded value for a friendly enum name.
-        """
-
-        options = self.value(key).options
-        options_inv = {v: k for k, v in options.items()}  # Invert the map.
-        return options_inv[name]
-
-    def enum_name(self, key, value):
-        """Look up the friendly enum name for an encoded value.
-        """
-
-        options = self.value(key).options
-        return options[value]
-
-    @property
-    def binary_monitor_data(self):
-        """Check that type of monitoring is BINARY(BYTE).
-        """
-
-        return self.data['Monitoring']['type'] == 'BINARY(BYTE)'
-
-    def decode_monitor_binary(self, data):
-        """Decode binary encoded status data.
-        """
-
-        decoded = {}
-        for item in self.data['Monitoring']['protocol']:
-            key = item['value']
-            value = 0
-            for v in data[item['startByte']:item['startByte'] +
-                          item['length']]:
-                value = (value << 8) + v
-            decoded[key] = str(value)
-        return decoded
-
-    def decode_monitor_json(self, data):
-        """Decode a bytestring that encodes JSON status data."""
-
-        return json.loads(data.decode('utf8'))
-
-    def decode_monitor(self, data):
-        """Decode  status data."""
-
-        if self.binary_monitor_data:
-            return self.decode_monitor_binary(data)
-        else:
-            return self.decode_monitor_json(data)
-
-
-class Device(object):
-    """A higher-level interface to a specific device.
-
-    Unlike `DeviceInfo`, which just stores data *about* a device,
-    `Device` objects refer to their client and can perform operations
-    regarding the device.
-    """
-
-    def __init__(self, client, device):
-        """Create a wrapper for a `DeviceInfo` object associated with a
-        `Client`.
-        """
-
-        self.client = client
-        self.device = device
-        self.model = client.model_info(device)
-
-    def _set_control(self, key, value):
-        """Set a device's control for `key` to `value`.
-        """
-
-        self.client.session.set_device_controls(
-            self.device.id,
-            {key: value},
-        )
-
-    def _get_config(self, key):
-        """Look up a device's configuration for a given value.
-
-        The response is parsed as base64-encoded JSON.
-        """
-
-        data = self.client.session.get_device_config(
-            self.device.id,
-            key,
-        )
-        return json.loads(base64.b64decode(data).decode('utf8'))
-
-    def _get_control(self, key):
-        """Look up a device's control value.
-        """
-
-        data = self.client.session.get_device_config(
-            self.device.id,
-            key,
-            'Control',
-        )
-
-        # The response comes in a funky key/value format: "(key:value)".
-        _, value = data[1:-1].split(':')
-        return value
-
-
-class ACMode(enum.Enum):
-    """The operation mode for an AC/HVAC device."""
-
-    COOL = "@AC_MAIN_OPERATION_MODE_COOL_W"
-    DRY = "@AC_MAIN_OPERATION_MODE_DRY_W"
-    FAN = "@AC_MAIN_OPERATION_MODE_FAN_W"
-    AI = "@AC_MAIN_OPERATION_MODE_AI_W"
-    HEAT = "@AC_MAIN_OPERATION_MODE_HEAT_W"
-    AIRCLEAN = "@AC_MAIN_OPERATION_MODE_AIRCLEAN_W"
-    ACO = "@AC_MAIN_OPERATION_MODE_ACO_W"
-    AROMA = "@AC_MAIN_OPERATION_MODE_AROMA_W"
-    ENERGY_SAVING = "@AC_MAIN_OPERATION_MODE_ENERGY_SAVING_W"
-
-
-class ACFanSpeed(enum.Enum):
-    """The fan speed for an AC/HVAC device."""
-
-    SLOW = '@AC_MAIN_WIND_STRENGTH_SLOW_W'
-    SLOW_LOW = '@AC_MAIN_WIND_STRENGTH_SLOW_LOW_W'
-    LOW = '@AC_MAIN_WIND_STRENGTH_LOW_W'
-    LOW_MID = '@AC_MAIN_WIND_STRENGTH_LOW_MID_W'
-    MID = '@AC_MAIN_WIND_STRENGTH_MID_W'
-    MID_HIGH = '@AC_MAIN_WIND_STRENGTH_MID_HIGH_W'
-    HIGH = '@AC_MAIN_WIND_STRENGTH_HIGH_W'
-    POWER = '@AC_MAIN_WIND_STRENGTH_POWER_W'
-    AUTO = '@AC_MAIN_WIND_STRENGTH_AUTO_W'
-
-
-class ACOp(enum.Enum):
-    """Whether a device is on or off."""
-
-    OFF = "@AC_MAIN_OPERATION_OFF_W"
-    RIGHT_ON = "@AC_MAIN_OPERATION_RIGHT_ON_W"  # This one seems to mean "on"?
-    LEFT_ON = "@AC_MAIN_OPERATION_LEFT_ON_W"
-    ALL_ON = "@AC_MAIN_OPERATION_ALL_ON_W"
-
-
-class ACDevice(Device):
-    """Higher-level operations on an AC/HVAC device, such as a heat
-    pump.
-    """
-
-    @property
-    def f2c(self):
-        """Get a dictionary mapping Fahrenheit to Celsius temperatures for
-        this device.
-
-        Unbelievably, SmartThinQ devices have their own lookup tables
-        for mapping the two temperature scales. You can get *close* by
-        using a real conversion between the two temperature scales, but
-        precise control requires using the custom LUT.
-        """
-
-        mapping = self.model.value('TempFahToCel').options
-        return {int(f): c for f, c in mapping.items()}
-
-    @property
-    def c2f(self):
-        """Get an inverse mapping from Celsius to Fahrenheit.
-
-        Just as unbelievably, this is not exactly the inverse of the
-        `f2c` map. There are a few values in this reverse mapping that
-        are not in the other.
-        """
-
-        mapping = self.model.value('TempCelToFah').options
-        out = {}
-        for c, f in mapping.items():
-            try:
-                c_num = int(c)
-            except ValueError:
-                c_num = float(c)
-            out[c_num] = f
-        return out
-
-    def set_celsius(self, c):
-        """Set the device's target temperature in Celsius degrees.
-        """
-
-        self._set_control('TempCfg', c)
-
-    def set_fahrenheit(self, f):
-        """Set the device's target temperature in Fahrenheit degrees.
-        """
-
-        self.set_celsius(self.f2c[f])
-
-    def set_zones(self, zones):
-        """Turn off or on the device's zones.
-
-        The `zones` parameter is a list of dicts with these keys:
-        - "No": The zone index. A string containing a number,
-          starting from 1.
-        - "Cfg": Whether the zone is enabled. A string, either "1" or
-          "0".
-        - "State": Whether the zone is open. Also "1" or "0".
-        """
-
-        # Ensure at least one zone is enabled: we can't turn all zones
-        # off simultaneously.
-        on_count = sum(int(zone['State']) for zone in zones)
-        if on_count > 0:
-            zone_cmd = '/'.join(
-                '{}_{}'.format(zone['No'], zone['State'])
-                for zone in zones if zone['Cfg'] == '1'
-            )
-            self._set_control('DuctZone', zone_cmd)
-
-    def get_zones(self):
-        """Get the status of the zones, including whether a zone is
-        configured.
-
-        The result is a list of dicts with the same format as described in
-        `set_zones`.
-        """
-
-        return self._get_config('DuctZone')
-
-    def set_fan_speed(self, speed):
-        """Set the fan speed to a value from the `ACFanSpeed` enum.
-        """
-
-        speed_value = self.model.enum_value('WindStrength', speed.value)
-        self._set_control('WindStrength', speed_value)
-
-    def set_mode(self, mode):
-        """Set the device's operating mode to an `OpMode` value.
-        """
-
-        mode_value = self.model.enum_value('OpMode', mode.value)
-        self._set_control('OpMode', mode_value)
-
-    def set_on(self, is_on):
-        """Turn on or off the device (according to a boolean).
-        """
-
-        op = ACOp.RIGHT_ON if is_on else ACOp.OFF
-        op_value = self.model.enum_value('Operation', op.value)
-        self._set_control('Operation', op_value)
-
-    def get_filter_state(self):
-        """Get information about the filter."""
-
-        return self._get_config('Filter')
-
-    def get_mfilter_state(self):
-        """Get information about the "MFilter" (not sure what this is).
-        """
-
-        return self._get_config('MFilter')
-
-    def get_energy_target(self):
-        """Get the configured energy target data."""
-
-        return self._get_config('EnergyDesiredValue')
-
-    def get_light(self):
-        """Get a Boolean indicating whether the display light is on."""
-
-        value = self._get_control('DisplayControl')
-        return value == '0'  # Seems backwards, but isn't.
-
-    def get_volume(self):
-        """Get the speaker volume level."""
-
-        value = self._get_control('SpkVolume')
-        return int(value)
-
-    def monitor_start(self):
-        """Start monitoring the device's status."""
-
-        mon = Monitor(self.client.session, self.device.id)
-        mon.start()
-        self.mon = mon
-
-    def monitor_stop(self):
-        """Stop monitoring the device's status."""
-
-        self.mon.stop()
-
-    def poll(self):
-        """Poll the device's current state.
-
-        Monitoring must be started first with `monitor_start`. Return
-        either an `ACStatus` object or `None` if the status is not yet
-        available.
-        """
-
-        # Abort if monitoring has not started yet.
-        if not hasattr(self, 'mon'):
-            return None
-
-        res = self.mon.poll_json()
-        if res:
-            return ACStatus(self, res)
-        else:
-            return None
-
-
-class ACStatus(object):
-    """Higher-level information about an AC device's current status.
-    """
-
-    def __init__(self, ac, data):
-        self.ac = ac
-        self.data = data
-
-    @staticmethod
-    def _str_to_num(s):
-        """Convert a string to either an `int` or a `float`.
-
-        Troublingly, the API likes values like "18", without a trailing
-        ".0", for whole numbers. So we use `int`s for integers and
-        `float`s for non-whole numbers.
-        """
-
-        f = float(s)
-        if f == int(f):
-            return int(f)
-        else:
-            return f
-
-    @property
-    def temp_cur_c(self):
-        return self._str_to_num(self.data['TempCur'])
-
-    @property
-    def temp_cur_f(self):
-        return self.ac.c2f[self.temp_cur_c]
-
-    @property
-    def temp_cfg_c(self):
-        return self._str_to_num(self.data['TempCfg'])
-
-    @property
-    def temp_cfg_f(self):
-        return self.ac.c2f[self.temp_cfg_c]
-
-    def lookup_enum(self, key):
-        return self.ac.model.enum_name(key, self.data[key])
-
-    @property
-    def mode(self):
-        return ACMode(self.lookup_enum('OpMode'))
-
-    @property
-    def fan_speed(self):
-        return ACFanSpeed(self.lookup_enum('WindStrength'))
-
-    @property
-    def is_on(self):
-        op = ACOp(self.lookup_enum('Operation'))
-        return op != ACOp.OFF
