+++ wideq/errors.py
@@ -1,0 +1,37 @@
+class APIError(Exception):
+    """An error reported by the API."""
+
+    def __init__(self, code, message):
+        self.code = code
+        self.message = message
+
+
+class NotLoggedInError(APIError):
+    """The session is not valid or expired."""
+
+    def __init__(self):
+        pass
+
+
+class NotConnectedError(APIError):
+    """The service can't contact the specified device."""
+
+    def __init__(self):
+        pass
+
+
+class TokenError(APIError):
+    """An authentication token was rejected."""
+
+    def __init__(self):
+        pass
+
+
+class MonitorError(APIError):
+    """Monitoring a device failed, possibly because the monitoring
+    session failed and needs to be restarted.
+    """
+
+    def __init__(self, device_id, code):
+        self.device_id = device_id
+        self.code = code
